last commit

+++ routes/v1/book.php
@@ -5,10 +5,10 @@ use Illuminate\Support\Facades\Route;
 
 
 Route::get('/', [BookController::class, 'index']);
+Route::get('{id}', [BookController::class, 'show']);
 
-Route::prefix('/')->middleware('auth:sanctum', 'role:admin')->group(function () {
-    Route::get('{id}', [BookController::class, 'show']);
-    Route::post('create', [BookController::class, 'store']);
-    Route::post('edit/{id}', [BookController::class, 'update']);
-    Route::delete('delete/{id}', [BookController::class, 'destroy']);
+Route::prefix('/')->middleware('auth:sanctum', 'role:admin|editor')->group(function () {
+    Route::post('create', [BookController::class, 'store'])->middleware('permission:create book');
+    Route::post('edit/{id}', [BookController::class, 'update'])->middleware('permission:edit book');
+    Route::delete('delete/{id}', [BookController::class, 'destroy'])->middleware('permission:delete author');
 });
