+++ routes/v1/category.php
@@ -8,8 +8,8 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', [CategoryController::class, 'index']);
 Route::get('{id}', [CategoryController::class, 'show']);
 
-Route::prefix('/')->middleware('auth:sanctum', 'role:admin')->group(function () {
-    Route::post('create', [CategoryController::class, 'store']);
-    Route::put('edit/{id}', [CategoryController::class, 'update']);
-    Route::delete('delete/{id}', [CategoryController::class, 'delete']);
+Route::prefix('/')->middleware('auth:sanctum', 'role:admin|editor')->group(function () {
+    Route::post('create', [CategoryController::class, 'store'])->middleware('permission:create category');
+    Route::put('edit/{id}', [CategoryController::class, 'update'])->middleware('permission:edit category');
+    Route::delete('delete/{id}', [CategoryController::class, 'delete'])->middleware('permission:delete category');
 });
